Turn CGTextbox from BG into sprite.
CGTextbox is now 7 sprites. This means that when CGTextbox displays, only BG3 is used, and only for the CG, meaning we can use all BGpalettes and the entirety of BGVRAM (minus BG0-BG3 screen entries) for 256-colour mode CGs.

OBJVRAM is also hardly used, so inserting a 72x32 sprite in there is fine. Same for OAM; 7 more sprites when only ~20 are used means we're still nowhere near the 128 sprite cap.