Bugfix nameless CGTEXTBOX.
When a CGTEXTBOXSTART is used to start text without a [SetName], the textbox wouldn't appear as its OAM was tied to the OAM for the namebox (which wouldn't be pushed to OAM). Made an extra hook to push a new nameboxless textbubble sprite to OAM instead.

Ch0's events have been extended to test more things.